补充cache

+++ ioc/config/cache/cache.go
@@ -19,7 +19,7 @@ type cache struct {
 	// 使用换成提供方, 默认使用GoCache提供的内存缓存
 	PROVIDER `json:"provider" yaml:"provider" toml:"provider" env:"CACHE_PROVIDER"`
 	// 单位秒, 默认5分钟
-	TTL int `json:"ttl" yaml:"ttl" toml:"ttl" env:"CACHE_TTL"`
+	TTL int64 `json:"ttl" yaml:"ttl" toml:"ttl" env:"CACHE_TTL"`
 
 	c Cache
 	ioc.ObjectImpl
